+++ app/models/bme.rb
@@ -41,7 +41,7 @@ class Bme < ActiveRecord::Base
     @bme_data=[]
     @bme_all.each do |bme|
       if !bme.measure_time.nil?
-        bme_data=[bme.measure_time.strftime("%Y-%m-%d %H:%M:%S").to_time.to_i*1000,bme.measure_value.to_i]
+        bme_data=[bme.measure_time.strftime("%Y-%m-%d %H:%M:%S").to_time.to_i*1000,bme.measure_value.to_f]
         @bme_data.append bme_data
       end
     end
